+++ crates/operp-settle/examples/export_batch.rs
@@ -65,7 +65,7 @@ fn main() {
     applied.push(tip);
     eng.ingest(d2).unwrap();
 
-    let px = 100_000 * PRICE_SCALE;
+    let px = 100_000 * PRICE_SCALE as i64;
     let qty = QTY_SCALE;
     let ask = sign_unit(
         vec![tip],
